requirements.txt
bcrypt==3.2.0
fastapi==0.63.0
passlib==1.7.4
pydantic==1.7.3
pytest==6.2.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
SQLAlchemy==1.4
starlette==0.13.6
uvicorn==0.13.3
